pwclient: Get rid of underscore to dash replacement
I was erroneously doing some string substitutions so that it
would find project names with an underscore in the name.  It
was only necessary due to a side effect of a bug fixed in commit
35241d528c3699b982730a333e236d57dc6451b6 (find project by linkname,
not full name).  In my test case, I had an underscore in the linkname,
but a dash in the full name.

No such string replacement is necessary, and actually breaks
filtering by projects with an underscore in the name now that
we're searching by linkname properly.

Signed-off-by: Nate Case <ncase@xes-inc.com>
Signed-off-by: Jeremy Kerr <jk@ozlabs.org>