[parser] Handle special case of "New newline at end of file"
If a file doesn't have a newline at the end, diff will insert a "\ No
newline at end of file" line in the middle of the last hunk. parser.py
was counting this line as part of the hunk line count. This had the side
effect of putting the last line of the hunk in the patch comment, and
truncating out the last line of the diff.

Handle this special case by not including this line in the internal line
counter. This appears to be the only case where diff will insert a
string like this within a hunk.

Signed-off-by: Nate Case <ncase@xes-inc.com>
Signed-off-by: Jeremy Kerr <jk@ozlabs.org>