Adjust postgres "grant all" script to new name of bundle table
Since c105cb0 (Initial bundle reordering support, 2009-01-29), the
patchwork_bundle_patches table and its associated id sequence does not
exist.  The script to grant appropriate privileges was not updated when
this change was made, and causes the set-up procedure described in
docs/INSTALL to fail.

This change adjust the script to grant appropriate access rights to
match the change in the DB schema.

Signed-off-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Jeremy Kerr <jk@ozlabs.org>