forms: Only pass post data to multiplepatchform if that was what was submitted
When updating bundles, we get a spurious error message in the
MultiplePatchForm; it's performing validation on the blank form data.

This change only passes the post data to the MuliplePatchForm if we
detect (through data['form']) that it was the MultiplePatchForm that was
submitted.

This fixes the current testsuite failure.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>