html: Turn the navbar into a breadcrumb bar
So we can have a somewhat of a hierarchical progression:

  All projects -> $project patches -> patch name

That should also help when we add series, we'll be able to just add the
series in there.

By default, we have:

  All projects -> $project patches

both being links, so we can return to the list of patches from any page
where the project is defined. The '$project patches' link becomes
insensitive in the list page. Finally the patch page adds the patch at
the end.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>
Signed-off-by: Belén Barros Peña <belen.barros.pena@intel.com>
Acked-by: Stephen Finucane <stephen.finucane@intel.com>