docs: Use 'prod' and 'dev' requirements files
Django 1.5 and 1.6 are no longer supported.

    https://docs.djangoproject.com/en/dev/internals/release-process/

As a result, one should not encourage people to develop and/or deploy
against these versions - the latest version of Django supported by
patchwork (currently 1.7) should be used for both.

In addition, rather than duplicating shared requirements for 'dev' and
'prod' environments, move all shared requirements to a "base" file.

Signed-off-by: Stephen Finucane <stephenfinucane@hotmail.com>
Signed-off-by: Jeremy Kerr <jk@ozlabs.org>