Update documentation and default settings to suit patchwork deployment model
We've always allowed configuration without altering any of the
version-controlled files. With the recent settings changes, we have an
extra level of indirection with the dev/prod settings modules.

Since we have to edit a config file anyway, this change moves the
prod.py settings file to a template, which is then used directly by
mange.py (and the wsgi application).

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>