patchwork: Explicitly load states fixtures
Apps with migrations will no longer load the initial_data fixtures by
default. In order to prepare to add migrations to patchwork, rename the
initial_data fixture to default_states (to match the default_tags
fixture), and explicitly load them in tests that require them.

Also, include this step in the INSTALL document.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>