Allow assigning of any user as delegate
Currently, Patchwork only allows to delegate patches to developers who
are registered as "maintainers" of the project in Patchwork or set up
as autodelegate "targets". This is a bit annoying as "maintainers" in
the Patchwork sense have the power to change the state of *any* patch.

Allow delegation of patches to developers who are not the submitter of
the patch in question, a maintainer of the project or an autodelegate
"target". This request is documented here:

    https://lists.ozlabs.org/pipermail/patchwork/2015-July/001351.html

Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>