tox: Remove ignored PEP8 errors and directories
Remove the following 'ignore' errors:

  E121-128,131  various visual indentation errors
  E251          spaces around keyword / parameter equals
  H405          multi-line docstring must have summary line

The only error remaining is E129, which is technically not part of the
PEP8 standard and contributes in practice besides lots of manual line
wrapping.

In addition, there's no good reason why the tests directory shouldn't
be assessed like everything else - tests are still code and, as such,
should be held to the same standards..

Finally, include the 'pwclient' script which should not be ignored.

Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>