Add support for 'django-debug-toolbar'
This tool is exceptionally helpful for debugging issues with Django:
install it as part of the 'dev' configuration. This only works on
Django > 1.6, due to a lack of support for older versions in the
upstream.

A note is included to help users avoid the issues seen when running
patchwork on a different machine.

Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>