docs: Reconfigure page structure
There are two distinct types of documentation available: documentation
for patchwork developers and documentation for patchwork deployers.
Rework the mkdocs configuration such that this hierarchy is
established.

As part of this change, the installation doc is renamed to deployment.
This guide should be used to install patchwork in a production
environment and not necessarily a deployment one. The existing name
was misleading so fix it.

Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>