requirements: Remove psycopg2
It would be good to provide the required packages for devs using both
MySQL and PostgreSQL. However, there does not appear to be any way to
specify an exclusive-or condition in pip requirements files and the
existing behavior of including dependencies for both RDBMSs means
system packages have to be installed for both, when you are likely
only using one. This is particularly problematic when you want to
set up a fast dev environment using something like Vagrant.

Settle on MySQL for now, until such a time as the consensus is to go
all-in with PostgreSQL (including docs, for example).

Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>